feat: [SKILL-6] migrate review_metrics.py into skill_bill/ package
Move the monolithic scripts/review_metrics.py (2475 lines) into a
proper Python package at skill_bill/ with 10 domain modules:
constants, db, config, review, triage, learnings, stats, sync,
output, cli.

Skills now call `skill-bill import-review` instead of resolving a
script path. install.sh uses `python3 -m skill_bill` for telemetry
setup. Zero external dependencies, all 71 tests pass.
