+++ .github/workflows/validate-agent-configs.yml
@@ -12,6 +12,14 @@ jobs:
       - name: Check out repository
         uses: actions/checkout@v4
 
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: '3.13'
+
+      - name: Install skill-bill package
+        run: pip install -e .
+
       - name: Set up Node.js
         uses: actions/setup-node@v4
         with:
