+++ skill_bill/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+__version__ = "0.1.0"
