+++ skill_bill/__main__.py
@@ -1,0 +1,3 @@
+from skill_bill.cli import main
+
+main()
